feat: Enhance privacy settings and refactor public family search
- Added 'Check All'/'Uncheck All' functionality to privacy settings in .
- Corrected TypeScript errors in .
- Updated English and Vietnamese locale files with new privacy-related translations.
- Refactored  to reuse existing specifications (, , ) and removed redundant custom specifications.
- Reverted  to its original state.
- Includes user's manual fixes for other related changes.
