+++ apps/admin/src/router/features/community.routes.ts
@@ -1,14 +1,21 @@
-import PublicFamilyListView from '@/views/community/PublicFamilyListView.vue';
+import { MainRouterView } from '@/views';
+import PublicFamilyListView from '@/views/family/PublicFamilyListView.vue';
 
 export const communityRoutes = [
   {
     path: '/community/families',
     name: 'PublicFamilyList',
-    component: PublicFamilyListView,
-    meta: {
-      requiresAuth: true,
-      title: 'menu.community', // Use i18n key for title
-      icon: 'mdi-account-group', // Add icon for the route
-    },
+    component: MainRouterView,
+    children: [
+      {
+      path: '',
+      component: PublicFamilyListView,
+      meta: {
+        requiresAuth: true,
+        title: 'menu.community', // Use i18n key for title
+        icon: 'mdi-account-group', // Add icon for the route
+      },
+    }
+    ]
   },
 ];
