+++ apps/admin/src/views/family/index.ts
@@ -2,4 +2,5 @@ export { default as FamilyAddView } from './FamilyAddView.vue';
 export { default as FamilyDetailView } from './FamilyDetailView.vue';
 export { default as FamilyEditView } from './FamilyEditView.vue';
 export { default as FamilyListView } from './FamilyListView.vue';
+export { default as PublicFamilyListView } from './PublicFamilyListView.vue';
